Fix MSVS 2010 build

+++ src/cpu/x64/jit_uni_reorder.cpp
@@ -924,10 +924,10 @@ struct jit_single_blk_kernel : public jit_generator {
                         mayiuse(avx512_core));
         if (!ok) return false;
 
-        ssize_t a = p.nodes[0].n;
+        int64_t a = p.nodes[0].n;
         auto b = p.nodes[0].is;
         auto c = p.nodes[0].os;
-        ssize_t d = p.nodes[1].n;
+        int64_t d = p.nodes[1].n;
         auto e = p.nodes[1].is;
         auto f = p.nodes[1].os;
